Deleted wrapperApp and remove BrowserRouter to index.js. Changed dependes in useEffect in file app.js. Changed actions and reducers in redux.
