+++ package.json
@@ -22,6 +22,7 @@
     "react-scripts": "5.0.1",
     "react-toastify": "^9.0.8",
     "redux": "^4.2.0",
+    "redux-logger": "^3.0.6",
     "redux-thunk": "^2.4.1",
     "reselect": "^4.1.6",
     "styled-components": "^5.3.5",
