+++ src/assets/styles/general.scss
@@ -1,5 +1,6 @@
 @import 'variables';
 @import "mixins";
+@import 'react-loading-skeleton/dist/skeleton.css';
 
 * {
   margin: 0;
