+++ src/constans/routes/index.js
@@ -1,0 +1,3 @@
+export const HOME = '/';
+export const PROFILE = '../profile';
+export const CONTACTS = '../contacts';
