--- src/sections/WrapperApp/styles.module.scss
@@ -1,10 +1,0 @@
-@import "../../assets/styles/variables";
-
-.wrapperApp {
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-  background: url('../../assets/images/mainBack.jpg') no-repeat center center, $white;
-  background-size: cover;
-  flex: auto;
-}
